+++ src/postgres/queries/delete_from.rs
@@ -1,9 +1,7 @@
 use std::fmt::{self, Display, Formatter};
 
-use itertools::Itertools;
-
 use crate::postgres::general::{Condition, OutputExpression, TableName, WithClause};
-use crate::tools::IntoIteratorOfSameType;
+use crate::tools::{joined, IntoIteratorOfSameType};
 
 /// Create a new `DELETE FROM` statement with the given table name.
 ///
@@ -117,11 +115,11 @@ impl Display for DeleteFrom {
         write!(f, "DELETE FROM {}", self.table_name,)?;
 
         if !self.where_.is_empty() {
-            write!(f, " WHERE {}", self.where_.iter().join(" AND "))?;
+            write!(f, " WHERE {}", joined(&self.where_, " AND "))?;
         }
 
         if !self.returning.is_empty() {
-            write!(f, " RETURNING {}", self.returning.iter().join(", "))?;
+            write!(f, " RETURNING {}", joined(&self.returning, ", "))?;
         }
 
         Ok(())