+++ src/tools.rs
@@ -1,7 +1,9 @@
 mod arrays;
+mod display;
 mod into_iterator_of_same_type;
 mod into_non_zero_array;
 
 pub use arrays::{build_array, transform_array};
+pub use display::joined;
 pub use into_iterator_of_same_type::IntoIteratorOfSameType;
 pub use into_non_zero_array::IntoNonZeroArray;